// src/types/payload-types.ts
export interface ShopPackage {
  id: number;
  name: string;
  description?: string | null;
  price: string;
  features?:
    | {
        feature: string;
        id?: string | null;
      }[]
    | null;
  whatsappLink?: string | null;
  updatedAt: string;
  createdAt: string;
}